requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
google-generativeai==0.3.2
openpyxl==3.1.2
gunicorn==22.0.0
